Add compilable unit game. I figured unit games would be easy but I'm not certain that I did it right.
Replaced the line:
  Definition unitTy := Wrapper [eqType of Unit] [finType of unit].
with:
  Definition unitTy := Wrapper [eqType of Unit] [finType of Unit].
I assume it was a typo but I don't know if there is some reason to use
"unit" (the compilable stuff would still build if you replace "mkUnit"
with "tt".

I had to undo the change from the "Fix so combinators builds again" to
get it to build. I don't know what that's about.