More work on big server step in INT'(WL) -> machine.v.
I spent a while trying to work with the rAllClientsSentCorrectly
defined in networkSemantics.v, and finally decided to just change
it. I had a feeling it would be hard to work with. The new definition
is in networkSemanticsNoBatch.v, and I've verified that it is still
implied by allClientsSentCorrectly as is required for the INT->INT'
simulation.