Use weightslang.state in the client package state again.
We can't use strict equality on client states between the network and
machine, since the oracle states aren't always in sync, so we just use
opto_oracle_eq and then relate the oracle state in the machine with
packets in the network.

There are just three lemmas remaining for the big server step and
they're somewhat similar to one another.