Finish client init step (except a little arithmetic lemma).
The proof is somewhat manual / repetitive in some places. I find it
hard to factor things nicely when it gets too complicated.

Besides the little arithmetic lemma (which I've verified is true using
Simp (⌐■_■)), only the client recv step remains which I expect to be
similar to init but maybe slightly harder.