Finish INT'(WL)->machine.v
The client recv step proof is hideous but technically done. There's a
lot of duplication between the two cases (SIter1 and SIterS). The
match relation preservation proof is pretty much the exact same for
both, so we should be able to factor it out into a separate lemma at
least.

I had to move the new iter semantics over to this branch, which broke
a few things. unifiedmachine.v is commented out entirely and some
proofs in machine.v are as well.